+++ packages/db/src/queries/index.ts
@@ -2,3 +2,5 @@ export * from "./workspaces";
 export * from "./projects";
 export * from "./tasks";
 export * from "./links";
+export * from "./members";
+export * from "./invites";