+++ packages/email/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./resend-adapter";
 export * from "./console-adapter";
+export * from "./templates/invite";